+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # `master`
 
+- 🚸 Add coloring for `Value::BatteryLife`
 - 💥 `Value::WindDirection` is dropped in favour of `Value::StringEnum`
 - 🚸 Notify when there is no data points for a chart
 - ♻️ Use `TryInto` to convert values for a chart